requirements.txt
pyOpenSSL
cryptography
validators
requests
rich
jinja2
PyYAML